refactor: split write.service tests into 3 files under 300 lines each
